#116: Bugfix - globalize saving and garbage collection logic
